Fixed routing and added snap to space cards
