+++ src/app/devices/index.js
@@ -1,40 +1,14 @@
-import { ScrollView, XStack, YStack } from "tamagui";
-import { View, StyleSheet } from "react-native";
-import StatCard from "@/components/statsCard";
-import ProductionCard from "@/components/productionCard";
-import SpaceCard from "@/components/spaces";
-import ProfilePic from "@/components/ProfilePic";
+import { StyleSheet, Text, View } from 'react-native'
+import React from 'react'
 
-export default function devices() {
+const index = () => {
   return (
-    <View style={styles.container}>
-      <ScrollView style={{ padding: 24 }}>
-        <YStack rowGap={10}>
-          <ProfilePic />
-          <ScrollView horizontal={true}>
-            <XStack>
-              <SpaceCard />
-              <SpaceCard />
-              <SpaceCard />
-            </XStack>
-          </ScrollView>
-
-          <ProductionCard />
-          <XStack columnGap={10}>
-            <StatCard />
-            <StatCard />
-          </XStack>
-        </YStack>
-      </ScrollView>
+    <View>
+      <Text>Devices</Text>
     </View>
-  );
+  )
 }
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: "#EFEFEF",
-    alignItems: "center",
-    justifyContent: "center",
-  },
-});
+export default index
+
+const styles = StyleSheet.create({})
